Optionally store vector-length in its header instead of a slot
This change is motivated by needing to eradicate assumptions from code
involving MAKE-ARRAY when called without an :INITIAL-something keyword,
whether heap or stack-allocated.
There is a new build feature flag (#+array-ubsan) which doesn't yet do
what it sounds like it does, but it will. For now, it avoids using the
LENGTH slot for its stated purpose, so that I can create a detector of
undefined read-before-write behavior by annotating all vectors at runtime
with 1 extra bit per element, without massively altering the representation.
The LENGTH slot will be repurposed to hold the ancillary data.

Ultimately we can leave this in as an option for users, depending on how
prevalent their bugs are. But if it comes to that, it may need a rethink
because this technique scales down ARRAY-DIMENSION-LIMIT by 2^32 which wasn't
a problem for my use-case, but could be a severe limitation for other people.

(Further note that with Memory Pool System, allocations are not prezeroed,
so SBCL internals may have assumptions about numeric arrays that become
false, but I don't yet know. I do know that some user code has problems)