Rename the NAME and UNTRUENAME slots in FILE-INFO.
The NAME slot is now TRUENAME, and UNTRUENAME is now PATHNAME. This
aligns the slots' names with the -PATHNAME*/-TRUENAME* variables and
the meaning of PATHNAME and TRUENAME for open file streams in
ANSI (the name used to open the file vs. the current canonical name
for the file, respectively).