Simplify all raw slot setting vops
* Remove the return value from raw slot setter vops, and presto!
  they're now pretty much equivalent to what the init vops were.
* Delete all the init vops
* Delete the "-C" variants where they existed

Some of the backends haven't specified that setters can accept an
immediate TN for the index. That's fixable.

Enforcing SETF semantics through use of vop result operands was a bad
idea, because it causes boxing of the return value (if used) since IR1
doesn't know that the vop returns its newval argument identically.
So even if the setter can't accept an immediate index, things are
better off this way.