Fix bug in CODE-HEADER-SET due to DEFINE-FULL-SETTER change
The CODE-HEADER-SET vop was expected to return a value.
This change removes that requirement, fixing the bug.

x86, ppc, and sparc were not problematic because the code-header vops
were not defined in terms of DEFINE-FULL-SETTER. The other architectures
got lucky because nothing that did a SETF of CODE-HEADER-REF actually
cared what the result was. Therefore omission of the final move was
not fatal despite being technically flawed.