Improve pathname handling in COMPILE-FILE
This fixes the single biggest problem with
(COMPILE-FILE "dir1/srcfile.lisp" :output-file "dir2/objfile")
by avoiding merging two relative directories into either
dir1/dir/ or dir2/dir1/ (and you don't know which)
Other lisp implementations disagreed with us, with each other,
and even with themselves about this. See the case studies in this
change as to why it's anyone's guess what it was supposed to do,
and how you can't even portably rely on COMPILE-FILE-PATHNAME
to not lie to you.

So, to gives users greater control, if :OUTPUT-FILE has a directory
in it, don't further manipulate it except to merge with *D-P-D*
which could always be bound to #P"". Users who want merging can merge
on their own. (You can't undo merging that was never wanted.)

And since the entire point of COMPILE-FILE-PATHNAME is to tell the
user what our implementation does, it seems fair game to implement
semantics that allow usage on modern content-addressable filesystems,
and make COMPILE-FILE-PATHNAME return the right answer instead of
a random hash based on the fasl file's sha256 sum or whatever.