WIP for post-freeze- Rewrite bit-bash fillers.
Not only was there the usual semantic problem of reading bytes despite only
trying to store bytes, which entails a performance hit in itself,
there is also a performance problem in the middle of the array where we
don't perform complicated masked writes, but nonetheless run slower than
whatever a C compiler and/or memset() would do for bulk operations.

So n the whole the new code is better, but in the A/B comparison it sometimes
has a row that looks worse. I think this is attributable to measurement noise,
because running the test again can cause the results to change,
and I didn't want to cherrypick the best of each row of from multiple runs.