Never source transform LIST or LIST* into CONS
Instead of having CONS be IR2-converted as FIXED-ALLOC + EMIT-INITS,
and two special cases of LIST or LIST* becoming CONS, have all three
go through the LIST* vop.
This allows more precise control over the scope of pseudo-atomic
in all cases, which lets us choose between either ending the PA section
as soon as 1 or 2 conses are made (if that's the total number to be made),
or initializing entirely inside the PA which will work better for removing
the assumption that memory is prezeroed.
(Otherwise fixed-alloc would probably have to redundantly zeroize)