Avoid redundant initialization when adjusting a vector
:INITIAL-ELEMENT if processed by DATA-VECTOR-FROM-INITS would store
into the entire new array even though a leading portion of the result
would get copied from the old contents. Also we lacked an assertion
about the copied range, which I discovered invertently by introducing
a bug. Furthermore, SHRINK-VECTOR for decreasing the size was nonsense,
because it was equivalent to SUBSEQ which is equivalent to REPLACE on
a new vector, which is just the same as for increasing the size.

Remove bogus "... or single dimensional array" comment.

Fix style-warnings due to ignored result of ADJUST-ARRAY in tests.