Unteach genesis about MAKE-MUTEX.
There are three instances where genesis was doing make-mutex:
* Doing *ALLOCATOR-MUTEX* at cold load time was wholly
unnecessary. Changing !DEFINE-LOAD-TIME-GLOBAL to
DEFINE-LOAD-TIME-GLOBAL worked straight away.
* We set *PACKAGE-GRAPH-LOCKed* in the PACKAGE-COLD-INIT function. Easy
enough.
* It turns out that *WORLD-LOCK* is now a misnomer. Its uses are now
localized to the classoid/layout machinery, so it can get thrown into
class.lisp and logically cold initialized there, since its first use
is in fact in that cold-init function, making the dependency very
obvious. The name should change in the future to reduce confusion as
well.

The advantage of this is that genesis now needs to know one less
non-fundamental object layout and that files which create mutexes are
no longer incompatible with block compilation (or other settings where
fopcompilation is off).