Initialiaze some specials in cold init instead.
This reverses the change that made these initialized by cold load. For
reasons explained previously (see commit removing
!DEFINE-LOAD-TIME-GLOBAL), it is necessary for some features that we
not rely on genesis to set the values of certain variables. However,
I've tried to organize the cold initializations a bit better. Also,
I'm going to try to mark cold initializations which are technically
not necessary for bootstrap but good for debugging with #+sb-devel. Hopefully, that will make cold init seem less cluttered.