Delete early-source-locations.
Install SOURCE-LOCATION as a source transformation, so that it is
robust and so that there is no need to define it in the cross-compiler
early.

We are using DEFINE-SOURCE-TRANSFORM here like
SB-XC:DEFINE-COMPILER-MACRO if that existed, but it is also good to
express the intent that rewriting the call should be mandatory.