Warn less about duplicate-named slots
If the non-direct duplicate slot's name is not exported from any
package, then it must be an internal implementation detail, and in
principle the implementor of the new duplicately-named slot cannot
know about that internal detail.  (This implementation is subsequent
to discussions on #sbcl, but the fact that something like this
should be necessary is strongly suggested by the explicit handling
of *PCL-PACKAGE* in the previous iteration of this warning.)

There are still code patterns that will cause this warning to trigger
when the user code is not at fault -- for example, when a slot is
internally customized in a project with per-file packages, with its
name exported from internal packages but not from any public API
package, a distinction without any corresponding language support.
However, the revision in ths commit will not warn in any new
situations, and the warning is removed in the moderately common
pattern of all-slot-names-internal with heavy use of mixins.

Fixes lp#1943559