Eliminate the LIST* vop
List allocations go through a custom IR2 converter which can pass whatever
it wants for the info arg, so pass a boolean for dotted or not.
Simplify the bodies by asserting in the converter that at least one cons cell
is allocated. Also the :policy option is irrelevant and can be removed.