Perform modular arithmetic on ASH with unknown shift direction.
If the integer fits into a word it can be used inside modular
arithmetic expressions.
E.g.
(logand #xFF (ash (the fixnum i) (the fixnum c)))

no longer has a full call to ASH.

Could be extended to arbitrary integers if the shift amount and the
logand mask only use the bits from the first word.

The VOPs are implemented only for ARM64 here.