delete-no-op-vops: make the tn-ref workaround dumber.
Don't copy the VOPs of the original TN-REFs because if it's another
MOVE and it gets deleted the copies won't get propagated any further.
Just create references to the TNs which will never go away, which is
not great but stops other code from thinking that there are no
references to them.