Load symbols from the explicit runtime module handle on Windows
instead of relying on GetProcAddress with a NULL pointer to find the
symbol address in the current running executable (which worked when
just running the runtime as SBCL.EXE). This makes finding
the symbols much more robust and explicit, and in particular allows libsbcl.dll to
work correctly because the runtime symbols will be linked in from the in-process
runtime module (libsbcl.dll) rather than the executable which loaded
the runtime as a shared library.

This almost seems like it was an accidental omission, because it seems
weird to rely on the behavior of GetProcAddress passed with a NULL
module handle, when we already calculated where the real module handle
we wanted is.