Ignore type mismatches hidden in inlined code.
(car (if x 1 x)) signals a helpful warning, but when (if x 1 x) is
inside an inlined function then it's less helpful. Or for things like
(1+ (position 1 (the list y))) saying that there's a NIL there
somewhere.
Would be helpful with macros too, but they make it harder to separate
original code from macroexpansion code.