x86-64: Explicitly specify use of temp-reg-tn as vop :temporary
For the time being, wire a temporary to r11 so that we don't cause the
vop to use more temporary registers. Then when everything is made explicit
the wired offset can be removed, thus freeing up r11 for use anywhere.

Also, improve IF-EQ so that it can compare register vs boxed constant.