x86-64: Remove bad convention for invoking alloc_list()
Don't use the lsb of SIZE as the list/non-list flag.
At some point it avoided having 2x2x2 routines (for list/non-list,
saving YMM regs or not, targeting R11 or a random register)
but those are all not considerations now.
I meant to do this in rev 88d078fe and forgot to.