Parse memory size option in all executable cores
It's highly dubious that an executable with :save-runtime-options did not permit
increasing the dynamic-space size. Therefore parse all memory-related options
wherever they are on the command-line. Users of self-contained Lisp applications
don't need to know about a strict (and unconventional) required argument order,
even though they may understand the concept of setting the memory size.
And/or, it may be hard to construct the command-line in the prescribed order.
As remarked in the code, this is unlikely to conflict with application-level
flags because of the long spellings of the memory-size options.