Implement NLX-ENTRY-SINGLE on all other platforms.
NLX-ENTRY-SINGLE is a great optimization that is easy to implement and
vm-independent, so implement it for all platforms.

This has the benefit of removing ugly vop-existsp calls everywhere.

This is untested on MIPS, PPC, and SPARC, but it works on RISC-V, and
those platforms behave exactly the same way as RISC-V for NLX-ENTRY,
so I'm pretty confident this will work in those places too.