Implement a third data stream in %CODE-FIXUPS
These new fixups will be for patching in GC write barriers at core startup.
In x86 the barrier will resemble something like this:
  MOV TEMP, EA
  SHR TEMP, GC-CARD-SHIFT
  AND TEMP, MAX-CARD-INDEX
where the immediate operand to AND depends on the dynamic-space-size.

This barrier is similar to one used in JVM's G1GC (decades ago anyway).
But since we don't JIT-compile everything, we have to rely on patching.
A saved core that is restarted with the same space size that it was saved
with will not incur any startup overhead.

The fixup list mechanism is getting a bit overblown but it's convenient and
simpler than adding a new slot to code blobs, which I tried and failed at.