ppc64: Fix bug in old->young pointer detection of untagged fdefns
scav_code_header() uses is_lisp_pointer() to decide whether it should look
at the generation of each code header constant, which meant that it missed
all untagged fdefn pointers. Somehow it almost worked, just by accident.
The reason is that we almost never refer to a younger fdefn, except for
compiling to memory and referring to an undefined function.
Creating all fdefns sooner than the code eliminates that vulnerability.