Re(re)name PHYSENV to ENVIRONMENT.
* It's not obvious at all what the adjective "physical" is supposed to
mean, to the uninitiated, as this adjective means nothing in all
standard references in the literature. (Not the "Lispy" term is an
understatement. It's not a term at all.) It's better to use a clear
spelled out name which does in fact describe the IR1 data structure
pretty well. Using the word "environment" like this *is* the standard
in most compilers.
* This change also clarifies things, because not every use of
"environment" was updated to be qualified with "physical", making
matters even more confusing.
* Remove some bitrotted assertions and a few confusing comments that
sought to explain the environment structure via
reverse-engineering. The environment structure is described pretty
well in the internals manual for Python and other places in the
source.