Rename VESTIGIAL-EXIT-CAST to something that makes sense.
VESTIGIAL-EXIT-CAST isn't a cast at all, and despite the name, it
doesn't have anything to do with vestigial exits. Rather, its sole
purpose is to help implement MULTIPLE-VALUE-PROG1. Rename the node and
document its purpose clearly and what exactly messes up without it.

Its kind of a kludge that this thing is still a CAST node by
inheritance. It would be better to make an abstract node type FILTER
which CAST and this new node type subclassed, because many of the
operations on CASTs only deal with its filter semantics.