Fix verify_range() again, if soft protection
The condition for whether a page is touched is that either the object base
address is touched (for non-vectors) or the element is touched (for vectors),
but rev bcef9153 made it too lenient because it said _any_ page in a contiguous
range could satisfy the "unprotected" check. That's true for the GC per se,
but it's not the best check, because exactly 1 card is the correct card to look
at when asserting the protection invariant.