Try to clean up verify_range() somewhat
The reason this function is so strange relative to other varieties of a
widetag-based 'switch' is that this tries to be two things at once:
- parsing objects on the heap
- a loop over pointer words in spaces such as TLS and binding stack
It's slightly hard to achieve those objectives and retain clarity,
and I think it may be better to have three separate functions:
one to verify a single pointer, one to iterate over objects,
and one to verify TLS and bindings, etc. But I retained the overall
shape, while strengthening the assertion about lining up at objects.