Do not use Genesis to reflect CTYPEs or set constant values.
This commit is composed of two main almost-independent-but-not-quite
changes.

1. Do not reflect host CTYPEs into cold load for bootstrap.

03606c9205e683caf3ce667bf1fadf9e215db410 wanted to remove the OAOO
violating idiom of

(defglobal xxx -1)
(!cold-init-forms
 (setq xxx ...))
in the type system and also remove a META-INFO hack. The way it
accomplished this was by teaching genesis how to reflect host CTYPEs
into the target, which requires genesis to know both about how CTYPEs
are internally represented and also causes complications with making
sure the interned types at cold load time are sane with respect to
cold init time type initializations.

However, recent build order wrangling (merging most of the type system
back into one file, for example) allows us to solve this in a much
simpler way: We introduce a macro !COLD-INIT-TIME-VALUE to avoid
having to write a bunch of once-use globals and the fact that most of
the type system initialization happens in one file makes ordering the
initialization of specifier types and ctypes pretty simple. As shown
by the line diff count, this simplifies cold load quite a bit, and
removes hacks for specifier-type to directly evaluate into a constant
just so that the dumper can hand off the value to genesis and removes
the whole LITERAL-CTYPE-... stuff.

Also, deduplicate some logic in array.lisp with respect to fetching
element types from widetags in the process, avoiding some dumping of
host CTYPEs at the same time.

2. Do not use coordination between the fopcompiler (essentially a
top-level form special caser/interceptor) and Genesis to set
hairy DEFCONSTANT values during the cold load.

This also means that we need to go back to dumping named constant
references as essentially (load-time-value (symbol-global-value
+constant+)).

The rationale behind this is that it is becoming clear that special
casing toplevel forms hinders using the normal compiler, such as for
block compilation or a future CMU CL style auto top level form
smasher. It also has to be done to enable the above CTYPE change,
since having DEFCONSTANT symbol values be set at cold load time means
that the value forms need to be cold loadable. It also has the same
issue of relying on host value reflection, which is not super clean
considering we loses compile time side effects, so constants defined
in target-only code can't use this technique anyway. Therefore, it is
simpler to just rely on normal compiler machinery to do DEFCONSTANT
bootstrapping. I am also hoping to solve EQness of constants not being
preserved by block compilation, which this is going to help, since
fasl ops can't get ordered with smashed up top level forms at all,
while it may be possible to do something with normal load-time-value
lambdas.

Of course, there are downsides, as now certain constant values are no
longer available until top-level forms get run. Usually, it's not much
of a problem: We add extra #. to anonymize those constants that need
it, which is compensated by the removal of many #. which were forced
by the previous DEFCONSTANT-EQX. We change BLT-COPIER-FOR-WIDETAG to
not even use any global constant at all by localizing the value into a
closure.