Use FIND-FDEFN instead of (INFO :FUNCTION :DEFINITION)
FIND-FDEFN uses the state of the running image, whereas (theoretically)
INFO inquires what the compiler knows about the global environment.
That is usually equivalent in SBCL, but it can be too limiting, because
we don't support the idea of INFO being a generic API to inquire about
info from any abstract provider (i.e. does asking for INFO use FIND-FDEFN
or vice-versa?) For now, FIND-FDEFN uses INFO.
But in particular it should be possible to remove FDEFNs from globaldb
and store them directly on symbols *without* *increasing* the size of
a symbol. (The details of how aren't important, yet)