Compress symbol-package to 2 bytes on 64-bit builds
Steal the highest bytes of the NAME word as a package table index.
It does work to have more than 65536 packages, but SYMBOL-PACKAGE
will have degraded performance if that happens.

This is conditioned on #+compact-symbol, which is equivalent to #+64-bit
except on RISCV where I didn't implement it. This is not the best name of
the feature, but the patch is based on an experiment of shortening all
symbols to 4 words with only a single data slot for any or all of the fdefn,
plist, or dbinfo, with various fancy ways of storing each permutation.
That patch ended up not being more space-efficient, so I scrapped it.

As the next step, the free slot will be repurposed for SYMBOL-FDEFN.