Disregard FILLER_WIDETAG in conservative_root_p()
If the only remaining objects on a non-large-object page are fillers,
they must not prevent freeing the page due to ambiguous stack words.
I tried to ignore fillers by changing the LOWTAG_FOR_WIDETAG table
to not assign any lowtag, but we need fillers to be some kind of headered
object so that Lisp code can compute the size in MAP-ALLOCATED-OBJECTS.
The next-best choice was to explicitly check for that widetag.
The alternative shorter variant of conservative_root_p does not need
an extra check because you'd never have a precise pointer to filler.