Speed up conservative_stack_scan
Make conservative_root_p *less* exact by inserting any plausible pointer
into pinned_objects.  99% of the time, the pointer is valid. We still have
to filter out the invalid ones, which can be done in one pass over all keys
of pinned_objects. To see why this is better, consider one page with
several pinned objects {A,B,C,D,...} near the physical end:
   ..... 100 other objects ... | A | B | C
The old code had to skip over 100 things to find A, then skip over 101
to find B, etc. The new way also has to skip over objects, but it is
no longer quadratic in the total number of objects on the page.