Make cold package definitions look more like real ones.
To make moving things between the cold load and warm load easier, it's
better to make the definition of target packages look more like real
DEFPACKAGEs. As WHN notes in PRINCIPLES, the package-data-list stuff
is better expressed with machinery that looks more like plain
DEFPACKAGE. We might even be able to make these packages be defined by
vanilla DEFPACKAGE if we express :reexport entries manually, since all
the functions needed to grovel DEFPACKAGE data are standardized.

This also removes the quirk of having to have export lists be trees in
the package data struct, in favor of the normal multiple :export
clause style.