Fix rare card marking bug
Too much and not enough sticky marks: Scan the registers, but not the stack.

The failure could only occur if the interrupted thread needed its sigcontext
scanned. The reason this seldom failed (and other tests don't exhibit a failure)
is that most regression tests run their actual work and GC on the same thread.