gencgc: Rename gencgc-card-bytes to gencgc-page-bytes
This prepares for allowing cards that are smaller then pages by some factor.
So for example we might have 2KiB per card but 64KiB per page.
All of the code that referred to the "card" size was actually talking about
the value of the minimum allocation quantum, a/k/a "page".  There's also
allocation_granularity acting as a multiplier, but I might remove that, since
it seems to cause worse performance by forcing find_freeish_pages to iterate
when it didn't really need to. (If a free page exists as all, we should
return it. Large objects do their own thing no matter what, so it's neither
better nor worse; and large object allocation is infrequent, relatively).
I would imagine that a good solution is that pages should be larger,
but we still want cards to be small; thus the distinction is very important.