win32: Protect against Command Line Options differing from argv.
In situations where they aren't the same (for example, if the runtime
is being embedded in another program), we'd like to not have the
system crash. The resulting *posix-argv* will not be accurate,
but *posix-argv* should not be queried when the runtime is embedded in
another program, as we can't in general integrate our options with the
application's.