Simplify LATE-PROCLAIM stuff.
Type system bootstrapping has been sufficiently worked out since 2004
so that now *QUEUED-PROCLAIMS* is always NIL on the host. Therefore,
enforce the constraint that we never proclaim any types before the
type system is initialized, thereby allowing us to simplify deferred
proclamations so we don't have to do any of this on the host.