Initialize an array with unbound-markers instead.
Instead of reflecting host ctypes into the array, we fix up the types
in the array in cold-init instead, making it so that cold load doesn't
need to know how to patch load time values into another constant if we
choose to reconstruct ctypes at load time instead of cold loading them
directly into the image.

The order-sensitivity is not an issue because ARRAY-ELEMENT-CTYPE
should not be called before the type system gets initialized (and
unknown types get fixed up).