Rewrite a bad lisp test as a less bad C test
The gc-region-pickup test could corrupt memory, because it did not zeroize
its code blobs, and failed to recognize that the first word of the raw bytes
of a code blob contains a jump table word count. If that count were randomly
nonzero, then trans_code() would scramble that many following words.