Move the wp_cleared bit out of the page table
and into the mark table, if using physical page protection.
By doing so, we ensure that WP and WP-cleared are always read
together. The long comment formerly in protect_page() goes away,
and the write fault handler is easier to comprehend.

Also we make available another bit in the page table for
more page types.

Note that the value of CARD_MARKED is changed from 0 to 1
for #-soft-card-marks. (This shouldn't break arm64+darwin,
but it wouldn't surprise me if it did.)