Do not coalesce named and anonymous constants.
As the long comment shows, this has very unintuitive and most likely
unintended behavior from the spec, as anonymous constants with no
make-load-form suddenly compile fine literally with spooky action at a
distance just by having a named constant reference floating around
somewhere else in the file.

Simplifying this gives us more consistent and reasonable semantics for
constant coalescing and simplifies FIND-CONSTANT. We are able to do
this because now FIND-CONSTANT is no longer responsible for dumping
everything under the sun, and we can remove VERY SUSPICIOUS code to
boot.

Also get rid of that metaspace only code as it pertained to dumping,
which FIND-CONSTANT no longer does.