Fix 14 year old typo messing up lexenv computations.
26bd73ecbd7af2473ff97bfbdc7eae9e39f54ba4 accidentally changed a LET*
to a LET. This made LEXENVs get computed wrong (specifically anything
with *handled-conditions* and *policy*, as the LET binding for the
LEXENV essentially became effectless), so that

(declaim (optimize debug)
         (muffle-conditions compiler-note))

(defun foo ()
  (declare (speed 3))
  (+ x y))

improperly signalled a condition.

This bug was hidden on debug = 1 because the fopcompiler, which is not
enabled on high debug, does hide the compiler note. (I believe because
the fopcompiler ignores lexenvs completely, and it must have mutated
NULL-LEXENV directly, thereby accidentally giving the next few tlfs
the right value for *HANDLED-CONDITIONS* via LEXENV inheritance,
instead of actually using *HANDLED-CONDITIONS*). I wonder how many
other bugs were lurking because of this oversight, and how many
workarounds were introduced in the meantime.