Store MUTEX-%OWNER as 'struct thread *', not SB-THREAD:THREAD
Making %OWNER an untagged pointer saves a nontrivial amount of GC time
in applications that use hundreds of thousands of mutexes, especially if
mutexes are saved into the core image (hence the oldest generation).
In fact, mutexes are almost always old relative to threads, so fewer tagged
pointers from them is strictly better. This change is most effective for
soft-card-marks as it elides the mark step, more so for #-sb-futex where
%owner is the determiner of state, and brute-forced in a retry loop.

The down side is that MUTEX-OWNER entails work to determine, but production
use of that is suspect, because the owner could always appear as NIL when
the mutex is owned - since %OWNER is set only after acquisition - or non-NIL
when un-owned, for the opposite reason.  #-sb-futex has similar anomalies.
These are not new. In any case the value is from a random time in recent past.

It's worth mentioning that Bordeaux-Threads does not expose OWNER because
almost no other mutex API has that. But OWNED-P is a reasonable inquiry.