Make (load ".fasl" :print t) do something on high debug.
The spec allows us to print whatever we want while fasloading, but
certainly printing

;; NIL
;; NIL
;; NIL

as would be the case with code compiled on higher debug (or otherwise
not fopcompiled) is just confusing. Therefore, extend FASL-INPUT to
keep track of whether printing is supposed to happen, so FOPs
themselves can choose what information to print. For now, print when a
function entry gets loaded, and weaken the test to just test for
non-emptiness. We could try to print the primary value of
tlf-equivalent forms as before by instrumenting FOP-FUNCALL and
FOP-FUNCALL-FOR-EFFECT, but that seems somewhat not worth it when it
isn't mandated to do so.