Define CHECK-DESIGNATOR as a function, not a macro.
Reduces code size, and the only difference otherwise was that the
macro precomputed a format string, but that doesn't matter because
that only happens in the (very) exceptional case.

Defining as a function also loosens build order constraints, and also
means sharing the function across the cross-compiler and the target,
as well as removing a manual push onto *!REMOVABLE-SYMBOLS*.

Also, one CHECK-DESIGNATOR call had already been using a quoted MACRO
argument already, which would've technically been wrong previously.