Rewrite MAKE-LOAD-FORM-SAVING-SLOTS optimization.
The main purpose of this change is to de-bolt-on-ify the mlfss
recognition optimization.
* Doing pattern matching over Lisp code is not really super robust,
and while in this case some kind of code recognition seems
unavoidable, we now do much better, mainly by directly comparing with
the result of MAKE-LOAD-FORM-SAVING-SLOTS when possible rather than
making assumptions about the shape of the code it produces (though we
still do that a little bit).
* Make dumping instances via MAKE-LOAD-FORM-SAVING-SLOTS first class
in the dumper, so we don't need to try and manipulate the fasl table
and dump fops outside of the fasl dumper. At the same time, use the
standard circularity machinery in the dumper in place of needing two
different fops and an auxilliary list of slot names to dump an
instance. Now, all that's needed to reconstruct an instance is a
sequence of objects and slot names interleaved, and one FOP-INSTANCE,
reducing consing, fasl size, and making loading more efficient.
* Rewrite MAKE-LOAD-FORM-SAVING-SLOTS and EMIT-MAKE-LOAD-FORMS and
COMPILE-MAKE-LOAD-INIT-FORMS to have less junk.
* Change BUILT-IN-CLASSOID to have a compatible representation on the
host and the target, since otherwise the dumper will get confused on
how to externalize the object.
* LOAD-FORM-IS-DEFAULT-MLFSS-P is still needed, but only for the sake
of the fopcompiler.