Remove STRUCTURE!OBJECT.
It's not necessary. The only thing it did was act as a sanity check
for making sure we don't try to dump random things from the host. But
we can't dump random things from the host without explicitly supplying
some kind of cross compiler load form method anyway, in which case the
check is pretty much redundant. Random types from the host will also
not work when FIND-CLASSOID is called on it for the most part. We do
have to make sure that some target number types, which are represented
as host structs, do not count as instances, as they will not get
loaded as instances on the target.

This also makes it so we can remove a DEF!STRUCT whose only purpose
was to include STRUCTURE!OBJECT.