Remove get_alloc_pointer()
alloc_pointer as a concept might have made sense for cheneygc, where it is the
address at which to next allocate. But gencgc has only a high water mark which
the "allocation pointer" would certainly be strictly less than.
And dynamic_space_free_pointer itself is redundant, so remove that too.