riscv: LRAs were harmed in the making of this change.
Now that cheneygc has been condemned, there's less reason to keep LRA
as a concept, since the generational collector can help pin on-stack
code. We remove roughly 1.4 MB of code out of the cold core just from
the removal of all the machinery needed to manipulate LRAs, since we
can just use machine call/return instructions, which will be much
better for real hardware in any case. (RISC-V specifies how to hint
the return address stack prediction hardware, which we weren't doing
at all.)

We do add a third variation on code pinning, which is more
conservative than ARM64 (since we lack precise backtracing due to lack
of :fp-and-pc-standard-save) but otherwise more precise than PPC64 and
MIPS. Theoretically it should be possible to do some more bookkeeping
to get the backtracing right for code pinning as on ARM64.