Change representation of SIMD-PACK types
Instead of keeping a list-as-set of the specialized SIMD element types
for a particular union of SIMD-PACKs, use a bit-vector-as-set, as this
is not sensitive to the implementation details of set operations.
Change the various bits of the type system (TYPEP and its transforms,
mapping of CTYPE to primitive type) to accommodate this.

Using a bit-vector, in combination with our direct reflection of host
CTYPE objects into the target image, means that genesis needs to be
taught how to dump bit-vectors.  (The implementation is I hope correct
but is only tested on 3-element bit-vectors on x86-64.)

This change causes src/code/x86-64-vm.lisp to cross-compile
bitwise-identically from SBCL and CCL hosts.