Fix more fallout from rev 2b3585478cbf
I'm not sure how to test this- if the ELFinating lisp runs the same core as
the core being elfinated, everthing just works. I tried compiling more junk in,
but was unable to craft a sufficiently altered core to elicit the bug that
came up in a real-world situation with a core having >100,000 code blobs.