Add a workaround for yet another mmap() problem
The mips machine @ gcc23.fsffrance can't run make-target-2 twice in a row,
not for the usual reason of our MIPS code being suspect, but instead,
as evident through use of ldb immediately upon completing coreparse, the
cold core file on disk was corrupted by the previous run.
I couldn't see any place that we'd lose the MAP_PRIVATE flag based on
byte order or word size, and "strace" proved that we're passing the
right flags. So the only conclusion is a busted kernel.
How nobody else has succumbed to it is beyond me.