Define two specializations of redblack tree
One variant accepts WORD as the key and has no payload, the other
(which already existed) has FIXNUM as the key, plus a value.

I intend to use this to speed up conservative stack scan - Lisp can
maintain the tree and GC can read it. The pure functional nature of
the tree makes it safe to observe in the stop-the-world handler,
so each thread can be responsible for finding its own code roots
(The handoff to GC of that root set could be accomplished via bits
in the code header or by leaving a word in the header to construct
a chain, maybe- not sure if the latter can be made lockfree)
In reality, pseudo-static code should go in a sorted vector, which
is an even faster and denser representation of the tree.