riscv: Annotate load/stores of all sizes.
I don't know why I only restricted it to the word size to begin
with. A memory address is a memory address. We shouldn't care about
the size of the operation to annotate it. Now we see
pseudo-atomic-bits properly annotated in the disassembly. (We use SH
and LH to frob it.)