Reconstruct the result of +- overflow from the carry/sign flags.
Don't need to encode two registers for the error break, the result can
be targeted to one of the input operands, as they don't need to be
preserved for the error. And one of the operands can be an immediate
without worrying how to encode an error break with it.