Fix bug with disabled page protection
The use-case is that when trying to track down a SIGSEGV-related crash
resulting from an "actual" fault (i.e. not the GC implementation),
it helps to eliminate our mprotect() calls. However, it turns out that
setting the #define could cause a problem as explained in the comment.

The default for ENABLE_PAGE_PROTECTION is 1, so this is a no-op change
for everybody except developers.