constraint: Join type constraints on control flow confluence.
If basic blocks B1 and B2 both transfer control into B3, constraint
propagation would totally forget the type constraints on a variable X
if the constraint was not exactly the same in both B1 and B2.

Now we handle control flow join of type constraints more precisely, by
unioning the actual types themselves and adding the new constraints to
the intersection of constraint sets, rather than just forgetting about
them.

There is some trickery to deal with the fact that EQL constriant
propagation is optimistic: We need to avoid getting tripped up by
that.