x86-64: never cons in GENERIC-NEGATE
most-negative-fixnum is the only value that can overflow.
The hack in immobile-space is undesirable but we don't currently support
boxed constants in asm routines (I think).

Fix a bug in SORT-INLINE-CONSTANTS that could result in putting 16-byte
and larger constants preceding the jump table, which would clobber things
if heap relocation were needed.