Reduce system call overhead for opening fasls
As specified, *LOAD-TRUENAME* is costly to compute, and the truename lookup
for such a strange edge case as 'file.lisp' being a symlink to 'other.fasl'
is not something we have to cater to. (It's not unarguably right per se)

Add new logic to delay all realpath() calls unless actually querying truename.
Hide it behind #-ansi-compliant-load-truename, with the default of #+.
The fast logic gets it down to one access(), two fstat(), and one open(),
instead of >40 syscalls for a file 20 levels down from the root.

LOAD's pathname-type defaulting still works, but some tests fail because ...
I don't know why. If we could fix those, I'd bet the faster code could be made
the default, except for pedants. Meanwhile, this patch remains unadvertised,
mainly for people who care to follow the discussion about it,
which seems to be limited to just myself.
(https://groups.google.com/g/sbcl-devel/c/reHPQjCUPVU/m/sUlKOAcpAgAJ)