Adjust hashing test for 32-bit
This is basically a kludge for a test that doesn't test much of anything.
In tinkering with it, I saw the test pass many times, but unbeknown to me
it was because of always getting un-nicely-aligned addresses for dynamic-space.
If you disable ASLR and get the default address of #x09000000, it fails.
Anyway the point is that there are better hash functions than SXHASH.