Improve COPY-STRUCTURE
Use a layout flag bit to test for absence of raw slots.
This allows (DEFSTRUCT SHIP (X 0 :TYPE FIXNUM) (Y 0 :TYPE FIXNUM))
to take the fast loop even though there is a 0 bitmap
since GC does not need to read any slot.

Iterate over LAYOUT-BITMAP, not DD-BITMAP which avoids needing
the descriptive metadata in cold-init and is probably better in general.

Remove the obsolete-instance test which does not seem
to be prescribed by the spec.

Completely avoid use of the bitmap for x86[-64].