Fix instances of loop repeat .. for ..  pattern
Strictly speaking the variable clauses introduced by `for` must
precede the main clause introduced by `repeat`.  By placing the repeat
after the for, the semantics of the loop are changed observably in the
presence of a `finally` clause, which will see the effect of one extra
step of the iteration variable (and, if the step has side-effects,
those side-effects happen one more time).