Don't save the whole lexical environment in inline expansions.
Just what's needed from declarations and stuff. Do this by using the
code-walker. We could probably avoid making the syntactic closure for
this case as well by just stuffing the declarations somewhere in the
body. This can be done if we really want to get rid of syntactic
closures entirely.