Generate smaller+faster EQUALP predicate for defstruct
Avoid use of the slot-access-transform. Though this doesn't type-check
the "unsafe" slots, answering the question of whether two structures are
EQUALP is generally unaffected.  There may be an edge case involving two
instances that have a type-incorrect but EQ value in such a slot.
This changes considers them EQUALP. We could bypass the custom comparator
in safety 3 if it really mattered, but we'd need two variants of EQUALP,
and the policy at any call site would dictate which to call.
Similar consideration pertains to slots with unbound-marker in them.