Stop creating an instance of T
T's layout has 0 bitmap - which I think is indisputable.
It was a getting a slot vector which became orphaned in GC
and would have crashed if you called STD-INSTANCE-SLOTS after that.
(If anyone wants to propose a better CLASS-PROTOTYPE, I'm all ears)
This probably happened around the time that class-init got moved into
make-host-2 and other things changed quite a bit.